Debug the native C++ module in Java.

+++ core/test/testcanvas.h
@@ -8,7 +8,7 @@ struct GiCanvas;
 struct TestCanvas {
     static void initRand(unsigned seed = 99999);
     static int randInt(int minv, int maxv);
-    static int randFloat(float minv, float maxv, float div = 10.f);
+    static float randFloat(float minv, float maxv, float div = 10.f);
 
     static void testRect(GiCanvas* canvas);
     static void testLine(GiCanvas* canvas);
